chore: rename sidetree-v2 to sidetree
