+++ packages/element-api/src/express/app.js
@@ -16,9 +16,9 @@ const onErrorResponse = require('../../src/express/onErrorResponse');
 app.use(cors({ origin: true }));
 app.use(express.json());
 if (process.env.NODE_ENV === 'testing') {
-  app.set('sidetree-v2', require('../services/sidetree-v2-test'));
+  app.set('sidetree', require('../services/sidetree-test'));
 } else {
-  app.set('sidetree-v2', require('../services/sidetree-v2'));
+  app.set('sidetree', require('../services/sidetree'));
 }
 
 app.options('*', cors({ origin: true }));
